Fix UT error and add e2e w4a16 testcase.
